+++ modules/Sources/Features/Root/RootDestination.swift
@@ -67,35 +67,13 @@ extension Root {
                 return .none
 
             case .deeplinkWarning(.gotItTapped):
-                //                let destination = state.destinationState.previousDestination ?? state.destinationState.destination
-                //                return .send(.destination(.updateDestination(destination)))
+                state = .initial
+                state.splashAppeared = true
                 state.tabsState.selectedTab = .send
                 state.tabsState.sendState.destination = .scanQR
                 return .send(.destination(.updateDestination(.tabs)))
                 
             case .destination(.deeplink(let url)):
-//                // get the latest synchronizer state
-//                let synchronizerStatus = sdkSynchronizer.latestState().syncStatus
-//
-//                // process the deeplink only if app is initialized and synchronizer synced
-//                guard state.appInitializationState == .initialized && synchronizerStatus == .upToDate else {
-//                    // TODO: [#370] There are many different states and edge cases we need to handle here
-//                    // (https://github.com/Electric-Coin-Company/zashi-ios/issues/370)
-//                    return .none
-//                }
-//                return .run { send in
-//                    do {
-//                        await send(
-//                            try await process(
-//                                url: url,
-//                                deeplink: deeplink,
-//                                derivationTool: derivationTool
-//                            )
-//                        )
-//                    } catch {
-//                        await send(.destination(.deeplinkFailed(url, error.toZcashError())))
-//                    }
-//                }
                 if let _ = uriParser.checkRP(url.absoluteString) {
                     // The deeplink is some zip321, we ignore it and let users know in a warning screen
                     return .send(.destination(.updateDestination(.deeplinkWarning)))
@@ -188,7 +166,7 @@ extension Root {
                 return .none
 
             case .tabs, .initialization, .onboarding, .updateStateAfterConfigUpdate, .alert, .phraseDisplay, .synchronizerStateChanged,
-                    .welcome, .binding, .resetZashiFailed, .resetZashiSucceeded, .debug, .walletConfigLoaded, .exportLogs, .confirmationDialog,
+                    .welcome, .binding, .resetZashiSDKFailed, .resetZashiSDKSucceeded, .resetZashiKeychainFailed, .resetZashiKeychainRequest, .resetZashiFinishProcessing, .resetZashiKeychainFailedWithCorruptedData, .debug, .walletConfigLoaded, .exportLogs, .confirmationDialog,
                     .notEnoughFreeSpace, .serverSetup, .serverSetupBindingUpdated, .batteryStateChanged, .cancelAllRunningEffects, .addressBookBinding, .addressBook, .addressBookContactBinding, .addressBookAccessGranted, .osStatusError:
                 return .none
             }