+++ CHANGELOG.md
@@ -6,6 +6,17 @@ directly impact users rather than highlighting other crucial architectural updat
 
 ## [Unreleased]
 
+### Added
+- Tap to enlarge the Keystone Animated QR code.
+- Automatic full brightness for the Keystone Animated QR code.
+
+### Updated
+- Keystone SDK version bumped with scan improvements.
+- Reset Zashi flow enhanced with retry logic and better error handling.
+
+### Fixed
+- The sending screen occasionaly reapeared and stayed on screen forever.
+
 ## 1.3.2 build 1 (2025-01-09)
 
 ### Fixed